pyproject.toml
[tool.poetry]
name = "geoldb"
version = "0.1.0"
description = ""
authors = ["alemishch <mishchenko.ad@phystech.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.13.3"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.0"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
uvicorn = "^0.32.0"
fastapi = "^0.115.4"
streamlit = "^1.39.0"
requests = "^2.32.3"
streamlit-aggrid = "^1.0.5"
python-multipart = "^0.0.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
